Round 2 baby
Fixed the route.ts bug
